+++ src/main/java/com/easypost/exception/General/InvalidParameterError.java
@@ -1,24 +1,25 @@
 package com.easypost.exception.General;
 
+import com.easypost.Constants;
 import com.easypost.exception.EasyPostException;
 
 public class InvalidParameterError extends EasyPostException{
     /**
      * InvalidParameterError constructor.
      *
-     * @param message the exception message
+     * @param parameterName the name of the invalid parameter
      */
-    public InvalidParameterError(final String message) {
-        this(message, null);
+    public InvalidParameterError(final String parameterName) {
+        super(String.format(Constants.ErrorMessages.INVALID_PARAMETER, parameterName));
     }
 
     /**
-     * constructor.
+     * InvalidParameterError constructor with nested exception.
      *
-     * @param message the exception message
-     * @param e the Throwable object
+     * @param parameterName the name of the invalid parameter
+     * @param e the nested exception
      */
-    public InvalidParameterError(final String message, Throwable e) {
-        super(message, e);
+    public InvalidParameterError(final String parameterName, Throwable e) {
+        super(String.format(Constants.ErrorMessages.INVALID_PARAMETER, parameterName), e);
     }
 }