+++ src/main/java/com/easypost/model/TimeInTransit.java
@@ -1,6 +1,5 @@
 package com.easypost.model;
 
-import com.easypost.Constants;
 import com.easypost.exception.EasyPostException;
 import com.easypost.exception.General.InvalidParameterError;
 import com.google.gson.annotations.SerializedName;
@@ -52,7 +51,7 @@ public final class TimeInTransit {
             case "percentile_99":
                 return this.percentile99;
             default:
-                throw new InvalidParameterError(String.format(Constants.ErrorMessages.INVALID_PARAMETER, "percentile"));
+                throw new InvalidParameterError("percentile");
         }
     }
 
@@ -80,7 +79,7 @@ public final class TimeInTransit {
             case Percentile99:
                 return this.percentile99;
             default:
-                throw new EasyPostException("Invalid SmartrateAccuracy enum value.");
+                throw new InvalidParameterError("accuracy");
         }
     }