+++ src/main/java/com/easypost/service/ReferralCustomerService.java
@@ -1,6 +1,7 @@
 package com.easypost.service;
 
 import com.easypost.Constants;
+import com.easypost.exception.API.EncodingError;
 import com.easypost.exception.EasyPostException;
 import com.easypost.exception.API.ExternalApiError;
 import com.easypost.http.Requestor;
@@ -13,6 +14,7 @@ import com.easypost.utils.Utilities;
 
 import java.io.BufferedReader;
 import java.io.DataOutputStream;
+import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.HttpURLConnection;
 import java.net.URL;
@@ -25,7 +27,7 @@ public class ReferralCustomerService {
 
     /**
      * ReferralCustomerService constructor.
-     * 
+     *
      * @param client The client object.
      */
     ReferralCustomerService(EasyPostClient client) {
@@ -63,8 +65,8 @@ public class ReferralCustomerService {
         params.put("email", email);
         wrappedParams.put("user", params);
 
-        String url = String.format("%s/%s/%s/%s", client.getApiBase(), client.getApiVersion(), "referral_customers",
-                userId);
+        String url =
+                String.format("%s/%s/%s/%s", client.getApiBase(), client.getApiVersion(), "referral_customers", userId);
 
         Requestor.request(RequestMethod.PUT, url, wrappedParams, ReferralCustomer.class, client);
     }
@@ -76,8 +78,7 @@ public class ReferralCustomerService {
      * @return ReferralCustomerCollection object.
      * @throws EasyPostException when the request fails.
      */
-    public ReferralCustomerCollection all(final Map<String, Object> params)
-            throws EasyPostException {
+    public ReferralCustomerCollection all(final Map<String, Object> params) throws EasyPostException {
         String url = String.format("%s/%s/%s", client.getApiBase(), client.getApiVersion(), "referral_customers");
 
         return Requestor.request(RequestMethod.GET, url, params, ReferralCustomerCollection.class, client);
@@ -93,11 +94,11 @@ public class ReferralCustomerService {
      * @param expirationYear  Expiration year of the credit card.
      * @param cvc             CVC of the credit card.
      * @return PaymentMethodObject object.
-     * @throws Exception when the request fails.
+     * @throws EasyPostException when the request fails.
      */
     public PaymentMethodObject addCreditCardToUser(final String referralApiKey, final String number,
-            final int expirationMonth,
-            final int expirationYear, final String cvc) throws Exception {
+                                                   final int expirationMonth, final int expirationYear,
+                                                   final String cvc) throws EasyPostException {
         return addCreditCardToUser(referralApiKey, number, expirationMonth, expirationYear, cvc,
                 PaymentMethod.Priority.PRIMARY);
     }
@@ -113,12 +114,12 @@ public class ReferralCustomerService {
      * @param cvc             CVC of the credit card.
      * @param priority        Priority of this credit card.
      * @return PaymentMethodObject object.
-     * @throws Exception when the request fails.
+     * @throws EasyPostException when the request fails.
      */
     public PaymentMethodObject addCreditCardToUser(final String referralApiKey, final String number,
-            final int expirationMonth,
-            final int expirationYear, final String cvc,
-            final PaymentMethod.Priority priority) throws Exception {
+                                                   final int expirationMonth, final int expirationYear,
+                                                   final String cvc, final PaymentMethod.Priority priority)
+            throws EasyPostException {
         String easypostStripeApiKey = retrieveEasypostStripeApiKey();
         String stripeToken;
 
@@ -138,10 +139,10 @@ public class ReferralCustomerService {
      * @throws EasyPostException when the request fails.
      */
     private String retrieveEasypostStripeApiKey() throws EasyPostException {
-        String url = String.format("%s/%s/%s", client.getApiBase(), client.getApiVersion(),
-                "partners/stripe_public_key");
-        @SuppressWarnings("unchecked")
-        Map<String, String> response = Requestor.request(RequestMethod.GET, url, null, Map.class, client);
+        String url =
+                String.format("%s/%s/%s", client.getApiBase(), client.getApiVersion(), "partners/stripe_public_key");
+        @SuppressWarnings ("unchecked") Map<String, String> response =
+                Requestor.request(RequestMethod.GET, url, null, Map.class, client);
 
         return response.getOrDefault("public_key", "");
     }
@@ -155,11 +156,12 @@ public class ReferralCustomerService {
      * @param cvc                  CVC of the credit card.
      * @param easypostStripeApiKey EasyPost Stripe API key.
      * @return Stripe token.
-     * @throws Exception when the request fails.
+     * @throws EncodingError         when the request details could not be encoded.
+     * @throws IOException           when the request fails.
      */
     private static String createStripeToken(final String number, final int expirationMonth, final int expirationYear,
-            final String cvc,
-            final String easypostStripeApiKey) throws Exception {
+                                            final String cvc, final String easypostStripeApiKey)
+            throws EncodingError, IOException {
         Map<String, String> params = new HashMap<>();
         params.put("number", number);
         params.put("exp_month", String.valueOf(expirationMonth));
@@ -200,8 +202,8 @@ public class ReferralCustomerService {
 
         String responseBody = response.toString();
 
-        @SuppressWarnings("unchecked")
-        Map<String, Object> responseMap = Constants.Http.GSON.fromJson(responseBody, Map.class);
+        @SuppressWarnings ("unchecked") Map<String, Object> responseMap =
+                Constants.Http.GSON.fromJson(responseBody, Map.class);
 
         return responseMap.get("id").toString();
     }
@@ -216,7 +218,7 @@ public class ReferralCustomerService {
      * @throws EasyPostException when the request fails.
      */
     private PaymentMethodObject createEasypostCreditCard(final String referralApiKey, final String stripeObjectId,
-            final String priority) throws EasyPostException {
+                                                         final String priority) throws EasyPostException {
         Map<String, Object> params = new HashMap<>();
         params.put("stripe_object_id", stripeObjectId);
         params.put("priority", priority);